frontend: Home/ClusterTable: Extract ClusterTable from Home
