+++ frontend/src/i18n/locales/de/translation.json
@@ -23,12 +23,12 @@
   "Active": "Aktiv",
   "Plugin": "",
   "In-cluster": "",
-  "Home": "Startseite",
-  "All Clusters": "Alle Cluster",
   "Name": "Name",
   "Origin": "",
   "Status": "Status",
   "Warnings": "",
+  "Home": "Startseite",
+  "All Clusters": "Alle Cluster",
   "Load cluster": "Cluster laden",
   "Something went wrong with cluster {{ cluster }}": "Mit dem Cluster ist etwas schief gelaufen {{ cluster }}",
   "Choose another cluster": "Wählen Sie einen anderen Cluster",
