+++ frontend/src/i18n/locales/en/translation.json
@@ -23,12 +23,12 @@
   "Active": "Active",
   "Plugin": "Plugin",
   "In-cluster": "In-cluster",
-  "Home": "Home",
-  "All Clusters": "All Clusters",
   "Name": "Name",
   "Origin": "Origin",
   "Status": "Status",
   "Warnings": "Warnings",
+  "Home": "Home",
+  "All Clusters": "All Clusters",
   "Load cluster": "Load cluster",
   "Something went wrong with cluster {{ cluster }}": "Something went wrong with cluster {{ cluster }}",
   "Choose another cluster": "Choose another cluster",
