+++ frontend/src/i18n/locales/es/translation.json
@@ -23,12 +23,12 @@
   "Active": "Activo",
   "Plugin": "Plugin",
   "In-cluster": "En-cluster",
-  "Home": "Inicio",
-  "All Clusters": "Todos los Clusters",
   "Name": "Nombre",
   "Origin": "Origen",
   "Status": "Estado",
   "Warnings": "Avisos",
+  "Home": "Inicio",
+  "All Clusters": "Todos los Clusters",
   "Load cluster": "Cargar cluster",
   "Something went wrong with cluster {{ cluster }}": "Algo ha fallado en el cluster {{ cluster }}",
   "Choose another cluster": "Elegir un otro cluster",
