+++ frontend/src/i18n/locales/fr/translation.json
@@ -23,12 +23,12 @@
   "Active": "Actif",
   "Plugin": "",
   "In-cluster": "",
-  "Home": "Accueil",
-  "All Clusters": "Tous les clusters",
   "Name": "Nom",
   "Origin": "",
   "Status": "Statut",
   "Warnings": "",
+  "Home": "Accueil",
+  "All Clusters": "Tous les clusters",
   "Load cluster": "Charger cluster",
   "Something went wrong with cluster {{ cluster }}": "Quelque chose s'est mal passé avec le cluster {{ cluster }}",
   "Choose another cluster": "Choisir un autre cluster",
