+++ frontend/src/i18n/locales/it/translation.json
@@ -23,12 +23,12 @@
   "Active": "Attivo",
   "Plugin": "Estensione",
   "In-cluster": "In-cluster",
-  "Home": "Home",
-  "All Clusters": "Tutti i cluster",
   "Name": "Nome",
   "Origin": "Origine",
   "Status": "Stato",
   "Warnings": "Avvertenze",
+  "Home": "Home",
+  "All Clusters": "Tutti i cluster",
   "Load cluster": "Carica cluster",
   "Something went wrong with cluster {{ cluster }}": "Qualcosa è andato storto con il cluster {{ cluster }}",
   "Choose another cluster": "Scegli un altro cluster",
