+++ frontend/src/i18n/locales/pt/translation.json
@@ -23,12 +23,12 @@
   "Active": "Activo",
   "Plugin": "Plugin",
   "In-cluster": "No cluster",
-  "Home": "Início",
-  "All Clusters": "Todos os Clusters",
   "Name": "Nome",
   "Origin": "Origem",
   "Status": "Estado",
   "Warnings": "Avisos",
+  "Home": "Início",
+  "All Clusters": "Todos os Clusters",
   "Load cluster": "Carregar cluster",
   "Something went wrong with cluster {{ cluster }}": "Algo falhou com o cluster {{ cluster }}",
   "Choose another cluster": "Escolha outro cluster",
