+++ frontend/src/i18n/locales/zh-tw/translation.json
@@ -23,12 +23,12 @@
   "Active": "活躍",
   "Plugin": "外掛",
   "In-cluster": "叢集內",
-  "Home": "首頁",
-  "All Clusters": "所有叢集",
   "Name": "名稱",
   "Origin": "來源",
   "Status": "狀態",
   "Warnings": "警告",
+  "Home": "首頁",
+  "All Clusters": "所有叢集",
   "Load cluster": "讀取叢集",
   "Something went wrong with cluster {{ cluster }}": "叢集 {{ cluster }} 出現問題",
   "Choose another cluster": "選擇其他叢集",
